requirements.txt
discord.py
python-dotenv
openai
requests
pynacl
gTTS